+++ example/android/app/src/main/java/com/moengage/sampleapp/SampleApplication.kt
@@ -6,10 +6,12 @@ import com.moengage.core.config.FcmConfig
 import com.moengage.core.config.LogConfig
 import com.moengage.core.config.NotificationConfig
 import com.moengage.core.config.PushKitConfig
+import com.moengage.core.model.AccountMeta
 import com.moengage.core.model.SdkState
 import com.moengage.flutter.MoEInitializer
 import com.moengage.pushbase.MoEPushHelper
 import io.flutter.app.FlutterApplication
+import com.moengage.inapp.MoEInAppHelper
 
 /**
  * @author Umang Chamaria
@@ -34,6 +36,7 @@ class SampleApplication : FlutterApplication() {
             .configurePushKit(PushKitConfig(true))
         MoEInitializer.initialiseDefaultInstance(applicationContext, moEngage, SdkState.ENABLED,true)
         // optional, required in-case notification customisation is required.
-        MoEPushHelper.getInstance().registerMessageListener(CustomPushListener())
+        MoEPushHelper.getInstance().registerMessageListener(CustomPushListener(AccountMeta("DAO6UGZ73D9RTK8B5W96TPYN")))
+        MoEInAppHelper.getInstance().enableActivityRegistrationOnResume()
     }
 }