+++ core/moengage_flutter/lib/src/moengage_flutter.dart
@@ -402,7 +402,7 @@ class MoEngageFlutter {
   /// be displayed. If position is not passed, it will take default position
   /// [MoEngageNudgePosition.any]
   /// Note: This API is available for Android/iOS platforms. Not supported in Web Platform
-  /// @since TODO: Add version
+  /// @since 7.0.0
   void showNudge({MoEngageNudgePosition position = MoEngageNudgePosition.any}) {
     _platform.showNudge(position, appId);
   }