+++ core/moengage_flutter_platform_interface/lib/src/model/inapp/nudge_position.dart
@@ -1,5 +1,5 @@
 /// Position for Non-Intrusive Nudge Campaigns
-/// @since TODO: Add Version
+/// @since 2.0.0
 enum MoEngageNudgePosition {
   /// For nudge to be shown at the TOP
   top,